gan ref

+++ data/d2/datasets/superpixels.py
@@ -11,15 +11,24 @@ from PIL import Image
 import cv2
 from torch import Tensor
 from torchvision.datasets import VisionDataset
-from superpixels.mbs import superpixels, proc_pool
+from superpixels.mbs import superpixels
 
 
 class Superpixels(VisionDataset):
 
-    def __init__(self, root, compute_sp=True, transform=None, target_transform=None, transforms=None, transforms_al=None):
+    def __init__(
+            self,
+            root,
+            target_root,
+            compute_sp=True,
+            transform=None,
+            target_transform=None,
+            transforms=None,
+            transforms_al=None
+    ):
         super(Superpixels, self).__init__(root, transforms, transform, target_transform)
         self.images_dir = self.root
-        self.targets_dir = 'cache/superpixels' + self.root
+        self.targets_dir = target_root
         self.images = []
         self.targets = []
         self.transforms_al = transforms_al
@@ -27,24 +36,26 @@ class Superpixels(VisionDataset):
         if not os.path.isdir(self.images_dir):
             raise RuntimeError('Dataset not found or incomplete.')
 
+        if not os.path.isdir(self.targets_dir) and compute_sp is False:
+            raise RuntimeError('Super pixels dataset not found or incomplete.')
+
         for file_name in os.listdir(self.images_dir):
             self.images.append(os.path.join(self.images_dir, file_name))
             self.targets.append(os.path.join(self.targets_dir, file_name) + ".npy")
 
-        def compute_superpixels_one(path):
-            img = cv2.imread(path)
+        def compute_superpixels_one(im_path, target_path):
+            img = cv2.imread(im_path)
             sp = superpixels(img)
-            sp_path = 'cache/superpixels' + path + ".npy"
-            np.save(sp_path, sp)
+            np.save(target_path, sp)
 
         if compute_sp:
             if os.path.exists(self.targets_dir):
                 shutil.rmtree(self.targets_dir)
             os.makedirs(self.targets_dir)
-            print("Compute superpixels ...")
-            for path in self.images:
-                compute_superpixels_one(path)
-            print("write into " + self.targets_dir)
+            print("Computing super pixels ...")
+            for i in range(len(self.images)):
+                compute_superpixels_one(self.images[i], self.targets[i])
+            print("written into " + self.targets_dir)
 
     def __getitem__(self, index) -> Tuple[Tensor, Tensor]:
